BookmarkChangeProcessor overrides now use size_t.
Chromium change:

https://chromium.googlesource.com/chromium/src/+/8cf4c23ebbd773240f207a8996941bd0343bd849

commit 8cf4c23ebbd773240f207a8996941bd0343bd849
Author: Peter Kasting <pkasting@chromium.org>
Date:   Thu Jun 6 00:38:30 2019 +0000

    Convert TreeNode{Model} Add/Remove functions to use size_t.

    This will make it more feasible to replace child_count() with children().size().

    This has a large number of followon effects.

    Bug: 956419

+++ components/brave_sync/client/bookmark_change_processor.h
@@ -71,19 +71,19 @@ class BookmarkChangeProcessor : public ChangeProcessor,
   void BookmarkModelBeingDeleted(bookmarks::BookmarkModel* model) override;
   void BookmarkNodeMoved(bookmarks::BookmarkModel* model,
                          const bookmarks::BookmarkNode* old_parent,
-                         int old_index,
+                         size_t old_index,
                          const bookmarks::BookmarkNode* new_parent,
-                         int new_index) override;
+                         size_t new_index) override;
   void BookmarkNodeAdded(bookmarks::BookmarkModel* model,
                          const bookmarks::BookmarkNode* parent,
-                         int index) override;
+                         size_t index) override;
   void OnWillRemoveBookmarks(bookmarks::BookmarkModel* model,
                              const bookmarks::BookmarkNode* parent,
-                             int old_index,
+                             size_t old_index,
                              const bookmarks::BookmarkNode* node) override;
   void BookmarkNodeRemoved(bookmarks::BookmarkModel* model,
                            const bookmarks::BookmarkNode* parent,
-                           int old_index,
+                           size_t old_index,
                            const bookmarks::BookmarkNode* node,
                            const std::set<GURL>& no_longer_bookmarked) override;
   void BookmarkAllUserNodesRemoved(bookmarks::BookmarkModel* model,
